I just updated the Git commit message editor. I have to make a change in order to have a commit message to type out.
I'm not sure how to save it. I see a checkmark in
the top right-hand corner. I'll see if there's a
keyboard shortcut for it.
I'm not sure, but it says "Accept Commit Message"
when I hover over it. So, I'll do that.

It promted me to save the changes first. --saved.